+++ packages/calcite-components-angular/projects/component-library/src/public-api.ts
@@ -1,3 +1,0 @@
-export * from "./lib/component-library.module";
-export { DIRECTIVES } from "./lib/stencil-generated";
-export * from "./lib/stencil-generated/components";